+++ crates/agentgateway/src/transport/mod.rs
@@ -3,11 +3,4 @@ mod rewind;
 pub mod stream;
 pub mod tls;
 
-#[derive(Debug, Clone)]
-pub struct BufferLimit(pub usize);
-
-impl BufferLimit {
-	pub fn new(limit: usize) -> Self {
-		BufferLimit(limit)
-	}
-}
+pub use agent_http::BufferLimit;
