+++ crates/http/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "agent-http"
+version = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+publish = { workspace = true }
+
+[lib]
+path = "src/lib.rs"
+
+[dependencies]
+axum.workspace = true
+axum-core.workspace = true
+bytes.workspace = true
+http.workspace = true
+
+[lints.rust]
+# This rule makes code more confusing
+mismatched_lifetime_syntaxes = "allow"
