+++ crates/llm/src/anthropic.rs
@@ -1,8 +1,7 @@
-use crate::apply;
 use agent_core::prelude::Strng;
 use agent_core::strng;
 
-use crate::RouteType;
+use crate::{RouteType, apply};
 
 #[apply(schema!)]
 #[cfg_attr(feature = "schema", schemars(rename = "AnthropicProvider"))]
