+++ crates/llm/src/conversion/rerank_tests.rs
@@ -1,11 +1,11 @@
-use crate::llm::types;
+use crate::types;
 
 // Request/response translation fidelity is covered by the golden snapshot tests in
 // `llm/tests.rs` (`requests/rerank/*` and `response/*/rerank*.json`); this module covers the
 // error and host-resolution paths.
 
-fn bedrock_provider(model: &str, region: &str) -> crate::llm::bedrock::Provider {
-	crate::llm::bedrock::Provider {
+fn bedrock_provider(model: &str, region: &str) -> crate::bedrock::Provider {
+	crate::bedrock::Provider {
 		model: Some(agent_core::strng::new(model)),
 		region: agent_core::strng::new(region),
 		guardrail_identifier: None,
@@ -19,7 +19,7 @@ fn test_bedrock_rerank_request_passes_through_full_arn() {
 	let req: types::rerank::Request =
 		serde_json::from_str(r#"{"query":"q","documents":["a"]}"#).unwrap();
 	let provider = bedrock_provider(arn, "us-east-1");
-	let out = crate::llm::conversion::bedrock::from_rerank::translate(&req, &provider).unwrap();
+	let out = crate::conversion::bedrock::from_rerank::translate(&req, &provider).unwrap();
 	let v: serde_json::Value = serde_json::from_slice(&out).unwrap();
 	assert_eq!(
 		v["rerankingConfiguration"]["bedrockRerankingConfiguration"]["modelConfiguration"]["modelArn"],
@@ -31,19 +31,17 @@ fn test_bedrock_rerank_request_passes_through_full_arn() {
 fn test_bedrock_rerank_uses_agent_runtime_host_and_rerank_path() {
 	let provider = bedrock_provider("cohere.rerank-v3-5:0", "us-west-2");
 	assert_eq!(
-		provider.get_host(crate::llm::RouteType::Rerank).as_str(),
+		provider.get_host(crate::RouteType::Rerank).as_str(),
 		"bedrock-agent-runtime.us-west-2.amazonaws.com"
 	);
 	assert_eq!(
 		provider
-			.get_path_for_route(crate::llm::RouteType::Rerank, false, "cohere.rerank-v3-5:0")
+			.get_path_for_route(crate::RouteType::Rerank, false, "cohere.rerank-v3-5:0")
 			.as_str(),
 		"/rerank"
 	);
 	assert_eq!(
-		provider
-			.get_host(crate::llm::RouteType::Embeddings)
-			.as_str(),
+		provider.get_host(crate::RouteType::Embeddings).as_str(),
 		"bedrock-runtime.us-west-2.amazonaws.com"
 	);
 }
@@ -52,7 +50,7 @@ fn test_bedrock_rerank_uses_agent_runtime_host_and_rerank_path() {
 /// rerank connection off `bedrock-runtime`.
 #[test]
 fn test_bedrock_connection_target_is_route_aware() {
-	use crate::llm::RouteType;
+	use crate::RouteType;
 
 	let provider = bedrock_provider("cohere.rerank-v3-5:0", "us-west-2");
 
@@ -75,8 +73,7 @@ fn test_bedrock_rerank_error_translation() {
 	let error_body = bytes::Bytes::from(
 		serde_json::to_vec(&serde_json::json!({"message": "model not found"})).unwrap(),
 	);
-	let translated =
-		crate::llm::conversion::bedrock::from_rerank::translate_error(&error_body).unwrap();
+	let translated = crate::conversion::bedrock::from_rerank::translate_error(&error_body).unwrap();
 	let error_resp: serde_json::Value = serde_json::from_slice(&translated).unwrap();
 	assert_eq!(error_resp["error"]["type"], "invalid_request_error");
 	assert_eq!(error_resp["error"]["message"], "model not found");
@@ -87,13 +84,13 @@ fn test_bedrock_rerank_empty_documents_errors() {
 	let req: types::rerank::Request =
 		serde_json::from_str(r#"{"query":"q","documents":[]}"#).unwrap();
 	let provider = bedrock_provider("cohere.rerank-v3-5:0", "us-east-1");
-	assert!(crate::llm::conversion::bedrock::from_rerank::translate(&req, &provider).is_err());
+	assert!(crate::conversion::bedrock::from_rerank::translate(&req, &provider).is_err());
 }
 
 // ---- Vertex Discovery Engine Ranking ----
 
-fn vertex_provider(project: &str, region: &str) -> crate::llm::vertex::Provider {
-	crate::llm::vertex::Provider {
+fn vertex_provider(project: &str, region: &str) -> crate::vertex::Provider {
+	crate::vertex::Provider {
 		model: None,
 		region: Some(agent_core::strng::new(region)),
 		project_id: agent_core::strng::new(project),
@@ -103,8 +100,7 @@ fn vertex_provider(project: &str, region: &str) -> crate::llm::vertex::Provider 
 #[test]
 fn test_vertex_rerank_response_rejects_non_numeric_id() {
 	let vertex_resp = r#"{"records":[{"id":"abc","score":0.5}]}"#;
-	let result =
-		crate::llm::conversion::vertex::from_rerank::translate_response(vertex_resp.as_bytes());
+	let result = crate::conversion::vertex::from_rerank::translate_response(vertex_resp.as_bytes());
 	assert!(result.is_err());
 }
 
@@ -112,7 +108,7 @@ fn test_vertex_rerank_response_rejects_non_numeric_id() {
 /// keeps the rerank connection off the `aiplatform` host.
 #[test]
 fn test_vertex_connection_target_is_route_aware() {
-	use crate::llm::RouteType;
+	use crate::RouteType;
 
 	let provider = vertex_provider("proj", "us-central1");
 
@@ -132,11 +128,11 @@ fn test_vertex_connection_target_is_route_aware() {
 
 #[test]
 fn test_vertex_rerank_uses_discovery_engine_host_and_ranking_path() {
-	use crate::llm::RouteType;
+	use crate::RouteType;
 
 	let provider = vertex_provider("proj", "global");
 	assert_eq!(
-		crate::llm::vertex::DISCOVERY_ENGINE_HOST.as_str(),
+		crate::vertex::DISCOVERY_ENGINE_HOST.as_str(),
 		"discoveryengine.googleapis.com"
 	);
 	let path = provider.get_path_for_model(RouteType::Rerank, None, false);
@@ -161,8 +157,7 @@ fn test_vertex_rerank_error_translation() {
 		}))
 		.unwrap(),
 	);
-	let translated =
-		crate::llm::conversion::vertex::from_rerank::translate_error(&error_body).unwrap();
+	let translated = crate::conversion::vertex::from_rerank::translate_error(&error_body).unwrap();
 	let error_resp: serde_json::Value = serde_json::from_slice(&translated).unwrap();
 	assert_eq!(error_resp["error"]["type"], "authentication_error");
 	assert_eq!(
@@ -176,5 +171,5 @@ fn test_vertex_rerank_empty_documents_errors() {
 	let req: types::rerank::Request =
 		serde_json::from_str(r#"{"query":"q","documents":[]}"#).unwrap();
 	let provider = vertex_provider("proj", "global");
-	assert!(crate::llm::conversion::vertex::from_rerank::translate(&req, &provider).is_err());
+	assert!(crate::conversion::vertex::from_rerank::translate(&req, &provider).is_err());
 }
