+++ crates/llm/src/custom.rs
@@ -1,8 +1,7 @@
-use crate::apply;
 use agent_core::prelude::Strng;
 use agent_core::strng;
 
-use crate::{InputFormat, RouteType};
+use crate::{InputFormat, RouteType, apply};
 
 #[apply(schema!)]
 #[cfg_attr(feature = "schema", schemars(rename = "CustomProvider"))]
