+++ crates/llm/src/vertex.rs
@@ -2,8 +2,7 @@ use agent_core::strng;
 use agent_core::strng::Strng;
 use serde_json::{Map, Value};
 
-use crate::apply;
-use crate::llm::{AIError, RouteType};
+use crate::{AIError, RouteType, apply};
 
 const ANTHROPIC_VERSION: &str = "vertex-2023-10-16";
